mgr/dashboard_v2: move datatable module
By moving the datatable module outside of the  components module we
can avoid a circular dependency between the datatable and sparkline
component.
