+++ src/pybind/mgr/dashboard_v2/frontend/src/app/shared/datatable/datatable.module.ts
@@ -4,7 +4,8 @@ import { FormsModule } from '@angular/forms';
 
 import { NgxDatatableModule } from '@swimlane/ngx-datatable';
 
-import { PipesModule } from '../../pipes/pipes.module';
+import { ComponentsModule } from '../components/components.module';
+import { PipesModule } from '../pipes/pipes.module';
 import { TableDetailsDirective } from './table-details.directive';
 import { TableKeyValueComponent } from './table-key-value/table-key-value.component';
 import { TableComponent } from './table/table.component';
@@ -14,7 +15,8 @@ import { TableComponent } from './table/table.component';
     CommonModule,
     NgxDatatableModule,
     FormsModule,
-    PipesModule
+    PipesModule,
+    ComponentsModule
   ],
   declarations: [
     TableComponent,
