+++ src/pybind/mgr/dashboard_v2/frontend/src/app/shared/datatable/table/table.component.html
