+++ src/pybind/mgr/dashboard_v2/frontend/src/app/shared/datatable/table/table.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../../defaults';
+@import '../../../../defaults';
 
 .dataTables_wrapper {
   margin-bottom: 25px;
