+++ src/pybind/mgr/dashboard_v2/frontend/src/app/shared/datatable/table/table.component.ts
@@ -15,7 +15,7 @@ import {
 import { DatatableComponent } from '@swimlane/ngx-datatable';
 import * as _ from 'lodash';
 
-import { CdTableColumn } from '../../../models/cd-table-column';
+import { CdTableColumn } from '../../models/cd-table-column';
 import { TableDetailsDirective } from '../table-details.directive';
 
 @Component({
